# docs/swagger.yaml
swagger: '2.0'
info:
  title: Living Progress - Build - Mentor Me
  description: Living Progress - Build - Mentor Me API.
  version: "1.0.0"
# the sample made up domain of the service
host: api.mentorme.com
...
schemes:
  - http
  - https
# will be prefixed to all paths
securityDefinitions:
  basicAuth:
    type: basic
# HTTP Basic Authentication. Works over `HTTP` and `HTTPS`
produces:
  - application/json
paths:
...
  /goals/{id}:
    get:
      tags:
        - Goal
      summary: get goal
      description: This endpoint is responsible for getting goal
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
      responses:
        200:
          description: The response containing the Goal
          schema:
            $ref: '#/definitions/Goal'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'

    put:
      tags:
        - Goal
      summary: update Goal
      description: This endpoint is responsible for updating Goal
      consumes:
        - multipart/form-data
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
        - name: subject
          in: formData
          description: The subject
          type: string
        - name: description
          in: formData
          description: The description
          type: string
        - name: number
          in: formData
          description: The number
          type: number
        - name: institutionalProgramId
          in: formData
          description: The institution program id
          type: number
        - name: goalCategory.id
          in: formData
          description: The goal category id
          type: number
        - name: durationInDays
          in: formData
          description: The duration in days
          type: number
        - name: custom
          in: formData
          description: The custom
          type: boolean
        - name: customData.mentor.id
          in: formData
          description: The custom mentor id
          type: number
        - name: customData.mentee.id
          in: formData
          description: The custom mentee id
          type: number
        - name: files
          in: formData
          description: The uploaded documents
          required: false
          type: file
      responses:
        200:
          description: The response containing the Goal
          schema:
            $ref: '#/definitions/Goal'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'

    delete:
      tags:
        - Goal
      summary: delete Goal
      description: This endpoint is responsible for deleting Goal
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
      responses:
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
  /goals:
    get:
      tags:
        - Goal
      summary: search for entities
      description: This endpoint is responsible for searching by criteria
      security:
      - basicAuth: []
      parameters:
         - name: institutionalProgramId
           in: query
           description: The institution program id
           type: number
           required: false
         - name: description
           in: query
           description: The description
           type: string
           required: false
         - name: subject
           in: query
           description: The subject
           type: string
           required: false
         - name: custom
           in: query
           description: The custom flag
           type: boolean
           required: false
         - name: pageNumber
           in: query
           description: The page number
           type: number
           format: int32
           required: false
         - name: pageSize
           in: query
           description: The page size
           type: number
           format: int32
           required: false
         - name: sortColumn
           in: query
           description: The sort column
           type: string
           required: false
         - name: sortOrder
           in: query
           description: The sort order
           type: string
           required: false
      responses:
        200:
          description: The response containing the InstitutionalPrograms
          schema:
            $ref: '#/definitions/GoalSearchResult'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - if token is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'

    post:
      tags:
        - Goal
      summary: create Goal
      description: This endpoint is responsible for creating Goal
      consumes:
        - multipart/form-data
      security:
        - basicAuth: []
      parameters:
        - name: subject
          in: formData
          description: The subject
          type: string
        - name: description
          in: formData
          description: The description
          type: string
        - name: number
          in: formData
          description: The number
          type: number
        - name: institutionalProgramId
          in: formData
          description: The institution program id
          type: number
        - name: goalCategory.id
          in: formData
          description: The goal category id
          type: number
        - name: durationInDays
          in: formData
          description: The duration in days
          type: number
        - name: custom
          in: formData
          description: The custom
          type: boolean
        - name: customData.mentor.id
          in: formData
          description: The custom mentor id
          type: number
        - name: customData.mentee.id
          in: formData
          description: The custom mentee id
          type: number
        - name: files
          in: formData
          description: The uploaded documents
          required: false
          type: file
      responses:
        200:
          description: The response containing the Goal
          schema:
            $ref: '#/definitions/Goal'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - if token is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  UserRole:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      value:
        type: string
        description: The value.
...
  Country:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      value:
        type: string
        description: The value.
  State:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      value:
        type: string
        description: The value.
...
  GoalCategory:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: The id.
        value:
          type: string
          description: The value.
...
  InstitutionContact:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: The id.
        title:
          type: string
          description: The title.
        firstName:
          type: string
          description: The first name.
        lastName:
          type: string
          description: The last name.
        email:
          type: string
          description: The email.
        phoneNumber:
          type: string
          description: The phone number.
        primaryContact:
          type: boolean
          description: The primary contact flag.
  ParentConsent:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: The id.
        parentName:
          type: integer
          format: int32
          description: The parent name.
        signatureFilePath:
          type: string
          description: The signature file path.
        parentEmail:
          type: string
          description: The parent email.
        token:
          type: string
          description: The token.
  InstitutionAffiliationCode:
        type: object
        properties:
          id:
            type: integer
            format: int32
            description: The id.
          code:
            type: string
            description: The code.
          institutionId:
            type: number
            description: The institution id.
          used:
            type: boolean
            description: The used flag.
  PersonalInterest:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: The id.
        value:
          type: string
          description: The value.
        picturePath:
          type: string
          description: The picture path.
        parentCategory:
          $ref: '#/definitions/PersonalInterest'
  WeightedPersonalInterest:
        type: object
        properties:
          id:
            type: integer
            format: int32
            description: The id.
          weight:
            type: number
            description: The weight.
          userId:
            type: number
            description: The user id.
          personalInterest:
            $ref: '#/definitions/PersonalInterest'
  ProfessionalInterest:
        type: object
        properties:
          id:
            type: integer
            format: int32
            description: The id.
          value:
            type: string
            description: The value.
          picturePath:
            type: string
            description: The picture path.
          parentCategory:
            $ref: '#/definitions/ProfessionalInterest'
  WeightedProfessionalInterest:
        type: object
        properties:
          id:
            type: integer
            format: int32
            description: The id.
          weight:
            type: number
            description: The weight.
          userId:
            type: number
            description: The user id.
          professionalInterest:
            $ref: '#/definitions/ProfessionalInterest'
...
  Institution:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      createdOn:
        type: integer
        format: int32
        description: The created on datetime with unix epoch format.
      institutionName:
        type: string
        description: The name.
      parentOrganization:
        type: string
        description: The parent organization.
      streetAddress:
        type: string
        description: The street address.
      city:
        type: string
        description: The city.
      zip:
        type: string
        description: The zip.
      phone:
         type: string
         description: The phone.
      email:
        type: string
        description: The email.
      description:
        type: string
        description: The description.
      status:
        type: string
        description: The status.
      logoPath:
        type: string
        description: The logo path.
      contacts:
        type: array
        items:
          $ref: '#/definitions/InstitutionContact'
      state:
        $ref: '#/definitions/State'
      country:
        $ref: '#/definitions/Country'
...
  Mentor:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      createdOn:
        type: integer
        format: int32
        description: The created on datetime with unix epoch format.
      username:
        type: string
        description: The username.
      firstName:
        type: string
        description: The first name.
      lastName:
        type: string
        description: The last name.
      email:
        type: string
        description: The email.
      profilePicturePath:
        type: string
        description: The profile picture path.
      status:
        type: string
        description: The status.
      assignedToInstitution:
        type: boolean
        description: The assigned to institution flag.
      birthDate:
        type: string
        description: The birth date.
      phone:
         type: string
         description: The phone.
      skypeUsername:
        type: string
        description: The skype username.
      introVideoLink:
        type: string
        description: The intro video link.
      description:
        type: string
        description: The description.
      averagePerformanceScore:
        type: number
        description: The average performance score.
      mentorType:
        type: string
        description: The mentor type.
      companyName:
        type: string
        description: The company name.
      linkedInUrl:
        type: string
        description: The linkedIn url.
      roles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
      personalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedPersonalInterest'
      professionalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedProfessionalInterest'
      institution:
        $ref: '#/definitions/Institution'
      institutionAffiliationCode:
        $ref: '#/definitions/InstitutionAffiliationCode'
      parentConsent:
        $ref: '#/definitions/ParentConsent'
...
  Mentee:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      createdOn:
        type: integer
        format: int32
        description: The created on datetime with unix epoch format.
      username:
        type: string
        description: The username.
      firstName:
        type: string
        description: The first name.
      lastName:
        type: string
        description: The last name.
      email:
        type: string
        description: The email.
      profilePicturePath:
        type: string
        description: The profile picture path.
      status:
        type: string
        description: The status.
      assignedToInstitution:
        type: boolean
        description: The assigned to institution flag.
      birthDate:
        type: string
        description: The birth date.
      phone:
         type: string
         description: The phone.
      skypeUsername:
        type: string
        description: The skype username.
      introVideoLink:
        type: string
        description: The intro video link.
      description:
        type: string
        description: The description.
      averagePerformanceScore:
        type: number
        description: The average performance score.
      familyIncome:
        type: number
        description: The family income.
      school:
        type: string
        description: The school.
      facebookUrl:
        type: string
        description: The facebook url.
      roles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
      personalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedPersonalInterest'
      professionalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedProfessionalInterest'
      institution:
        $ref: '#/definitions/Institution'
      institutionAffiliationCode:
        $ref: '#/definitions/InstitutionAffiliationCode'
      parentConsent:
        $ref: '#/definitions/ParentConsent'
...
  UsefulLink:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: The id.
        title:
          type: string
          description: The title.
        address:
          type: string
          description: The address.
  Document:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: The id.
        name:
          type: string
          description: The name.
        path:
          type: string
          description: The path.
        createdOn:
          type: number
          description: The created on time
        createdBy:
          type: number
          description: The created by user id.
  CustomAssignedGoalData:
      type: object
      properties:
        mentor:
          $ref: '#/definitions/Mentor'
        mentee:
          $ref: '#/definitions/Mentee'
  Goal:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      number:
        type: integer
        description: The number.
      subject:
        type: string
        description: The subject.
      description:
        type: string
        description: The description.
      durationInDays:
        type: integer
        description: The duration in days.
      custom:
        type: boolean
        description: The custom flag.
      institutionalProgramId:
        type: integer
        description: The institutional program id.
      goalCategory:
        $ref: '#/definitions/GoalCategory'
      customData:
        $ref: '#/definitions/CustomAssignedGoalData'
      tasks:
        type: array
        items:
          $ref: '#/definitions/Task'
      usefulLinks:
        type: array
        items:
         $ref: '#/definitions/UsefulLink'
      documents:
        type: array
        items:
          $ref: '#/definitions/Document'
...
  GoalSearchResult:
    type: object
    properties:
      total:
        type: number
        format: integer
        description: The total.
      totalPages:
        type: number
        format: integer
        description: The total pages.
      entities:
        type: array
        items:
          $ref: '#/definitions/Goal'
  CustomAssignedTaskData:
    type: object
    properties:
      mentor:
        $ref: '#/definitions/Mentor'
      mentee:
        $ref: '#/definitions/Mentee'
  Task:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      number:
        type: integer
        description: The number.
      description:
        type: string
        description: The description.
      durationInDays:
        type: integer
        description: The duration in days.
      custom:
        type: boolean
        description: The custom flag.
      mentorAssignment:
        type: boolean
        description: The mentor assignment flag.
      menteeAssignment:
        type: boolean
        description: The mentee assignment flag.
      goalId:
        type: integer
        description: The goal id.
      customData:
        $ref: '#/definitions/CustomAssignedTaskData'
      usefulLinks:
        type: array
        items:
         $ref: '#/definitions/UsefulLink'
      documents:
        type: array
        items:
          $ref: '#/definitions/Document'
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string